1. 整理代码目录 2. 增加了许多功能，大致的框架搭起来了

+++ src/chelper/lexer/LexerPos.cpp
@@ -4,9 +4,9 @@
 
 #include "LexerPos.h"
 
-CHelper::LexerPos::LexerPos(int line,
-                            int col,
-                            int which,
+CHelper::LexerPos::LexerPos(size_t line,
+                            size_t col,
+                            size_t which,
                             const std::string &filePath)
         : line(line),
           col(col),
