+++ src/chelper/lexer/LexerPos.h
@@ -11,12 +11,12 @@ namespace CHelper {
 
     class LexerPos {
     public:
-        int line; //这个文件的第几行
-        int col; //这一行的第几个字符
-        int index; //这个文件的第几个字符
+        size_t line; //这个文件的第几行
+        size_t col; //这一行的第几个字符
+        size_t index; //这个文件的第几个字符
         const std::string &filePath; //文件相对于functions文件夹的相对路径
 
-        LexerPos(int line, int col, int which, const std::string &fileName);
+        LexerPos(size_t line, size_t col, size_t which, const std::string &fileName);
 
         void next(char ch);
     };
